增加升级可忽略文件, 日志详细化 交前端测试 version3.0

+++ deploy_executable.py
@@ -4,17 +4,29 @@ import logging
 
 dirs = ["h5", "h5php", "sbin"]
 executables = ["zops_agentd", "zops_proxy", "zops_server"]
+ignore_subpaths = ["h5php/conf"]
+
+def should_ignore(path):
+    """检查路径是否应该被忽略"""
+    for subpath in ignore_subpaths:
+        if subpath in path:
+            logging.info(f"忽略 {path}", extra={"code": "200"})
+            return True
+    return False
 
 def delete_file_or_directory(path):
     """删除文件或目录"""
-    if os.path.exists(path):
+    if os.path.exists(path) and not should_ignore(path):
         try:
             if os.path.isfile(path):
                 os.remove(path)
                 logging.info(f"成功删除 {path}", extra={"code": "200"})
             else:
-                shutil.rmtree(path)
-                logging.info(f"成功删除 {path}", extra={"code": "200"})
+                for item in os.listdir(path):
+                    delete_file_or_directory(os.path.join(path, item))
+                if not os.listdir(path):  # 如果目录为空，则删除它
+                    os.rmdir(path)
+                    logging.info(f"成功删除 {path}", extra={"code": "200"})
         except OSError as e:
             logging.error(f"删除 {path} 失败: {e}", extra={"code": "500"})
             return False
@@ -22,13 +34,16 @@ def delete_file_or_directory(path):
 
 def copy(src, dst):
     """复制文件或目录"""
-    if os.path.exists(src):
+    if os.path.exists(src) and not should_ignore(src):
         try:
             if os.path.isfile(src):
                 shutil.copy2(src, dst)
                 logging.info(f"成功复制 {src} 到 {dst}", extra={"code": "200"})
             else:
-                shutil.copytree(src, dst)
+                if not os.path.exists(dst):
+                    os.makedirs(dst)
+                for item in os.listdir(src):
+                    copy(os.path.join(src, item), os.path.join(dst, item))
                 logging.info(f"成功复制 {src} 到 {dst}", extra={"code": "200"})
         except OSError as e:
             logging.error(f"复制 {src} 到 {dst} 失败: {e}", extra={"code": "500"})
